chore: reworked the other models
