+++ src/main/java/eu/tornplayground/tornapi/models/faction/basic/RaidWars.java
@@ -4,10 +4,14 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import eu.tornplayground.tornapi.models.Model;
 import eu.tornplayground.tornapi.models.converters.EpochLocalDateTimeConverter;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
 import java.util.Objects;
 
+@Getter
+@NoArgsConstructor(access = lombok.AccessLevel.PROTECTED)
 public class RaidWars extends Model {
     @JsonProperty("raiding_faction")
     private int raidingFaction;
@@ -24,29 +28,6 @@ public class RaidWars extends Model {
     @JsonProperty("start_time")
     @JsonDeserialize(converter = EpochLocalDateTimeConverter.class)
     private LocalDateTime startTime;
-
-    protected RaidWars() {
-    }
-
-    public int getRaidingFaction() {
-        return raidingFaction;
-    }
-
-    public int getDefendingFaction() {
-        return defendingFaction;
-    }
-
-    public float getRaiderScore() {
-        return raiderScore;
-    }
-
-    public int getDefenderScore() {
-        return defenderScore;
-    }
-
-    public LocalDateTime getStartTime() {
-        return startTime;
-    }
 
     @Override
     public boolean equals(Object o) {
