+++ src/main/java/eu/tornplayground/tornapi/models/torn/PawnShop.java
@@ -2,31 +2,20 @@ package eu.tornplayground.tornapi.models.torn;
 
 import eu.tornplayground.tornapi.models.Model;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import java.util.Objects;
 
+@Getter
+@NoArgsConstructor(access = lombok.AccessLevel.PROTECTED)
 public class PawnShop extends Model {
 
     @JsonProperty("points_value")
     private long pointsValue;
+
     @JsonProperty("donatorpack_value")
     private long donatorPackValue;
-
-    public long getPointsValue() {
-        return pointsValue;
-    }
-
-    public void setPointsValue(long pointsValue) {
-        this.pointsValue = pointsValue;
-    }
-
-    public long getDonatorPackValue() {
-        return donatorPackValue;
-    }
-
-    public void setDonatorPackValue(long donatorPackValue) {
-        this.donatorPackValue = donatorPackValue;
-    }
 
     @Override
     public boolean equals(Object o) {
