+++ src/main/java/eu/tornplayground/tornapi/models/torn/TornItem.java
@@ -2,114 +2,47 @@ package eu.tornplayground.tornapi.models.torn;
 
 import eu.tornplayground.tornapi.models.Model;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import java.util.Objects;
 
+@Getter
+@NoArgsConstructor(access = lombok.AccessLevel.PROTECTED)
 public class TornItem extends Model {
 
+    @JsonProperty("name")
     private String name;
+
+    @JsonProperty("description")
     private String description;
+
+    @JsonProperty("effect")
     private String effect;
+
+    @JsonProperty("requirement")
     private String requirement;
+
+    @JsonProperty("type")
     private String type;
+
     @JsonProperty("weapon_type")
     private String weaponType;
+
     @JsonProperty("buy_price")
     private long buyPrice;
+
     @JsonProperty("sell_price")
     private long sellPrice;
+
     @JsonProperty("market_value")
     private long marketValue;
+
+    @JsonProperty("circulation")
     private long circulation;
+
+    @JsonProperty("image")
     private String image;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getEffect() {
-        return effect;
-    }
-
-    public void setEffect(String effect) {
-        this.effect = effect;
-    }
-
-    public String getRequirement() {
-        return requirement;
-    }
-
-    public void setRequirement(String requirement) {
-        this.requirement = requirement;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public String getWeaponType() {
-        return weaponType;
-    }
-
-    public void setWeaponType(String weaponType) {
-        this.weaponType = weaponType;
-    }
-
-    public long getBuyPrice() {
-        return buyPrice;
-    }
-
-    public void setBuyPrice(long buyPrice) {
-        this.buyPrice = buyPrice;
-    }
-
-    public long getSellPrice() {
-        return sellPrice;
-    }
-
-    public void setSellPrice(long sellPrice) {
-        this.sellPrice = sellPrice;
-    }
-
-    public long getMarketValue() {
-        return marketValue;
-    }
-
-    public void setMarketValue(long marketValue) {
-        this.marketValue = marketValue;
-    }
-
-    public long getCirculation() {
-        return circulation;
-    }
-
-    public void setCirculation(long circulation) {
-        this.circulation = circulation;
-    }
-
-    public String getImage() {
-        return image;
-    }
-
-    public void setImage(String image) {
-        this.image = image;
-    }
 
     @Override
     public boolean equals(Object o) {
